+++ src/ImageViewer/docs/index.story.js
@@ -13,6 +13,11 @@ const exampleImageUrl = [
     value:
       'https://static.wixstatic.com/media/c78d05b79ede429fb77c9d8ec4443b93.jpg/v1/fit/w_375,h_375/c78d05b79ede429fb77c9d8ec4443b93.jpg',
   },
+  {
+    label: 'Small Image',
+    value:
+      'https://static.wixstatic.com/media/7175d1_bc0f6b3c4f43485aa99f0bcceaaea740~mv2.png',
+  },
 ];
 
 export default {